Labs/practice_07_pipeline/requirements.txt
torch 
torchvision
ultralytics
opencv-python
fastapi
uvicorn
psycopg2-binary
pydantic
